.dev/requirements.txt
qrcode
pyyaml
